requirements.txt
streamlit>=1.28.0
requests>=2.28.0
pandas>=1.5.0
PyPDF2>=3.0.0
python-pptx>=0.6.21
google-generativeai>=0.3.0
dataclasses; python_version < "3.7"
